Extract default GUI audio and sample browser views
